requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
streamlit==1.28.0
plotly==5.18.0
pyyaml==6.0.1
tqdm==4.66.1
feather-format==0.4.1